+++ src/classifiers.js
@@ -6,16 +6,16 @@
 
 const {hasComponentTag, hasHOCTag} = require('./utils');
 
-function defaultTypeClassifier ({section}) {
+async function defaultTypeClassifier ({section}) {
 	const kind = section.kind;
 
 	// Check for HOC
-	if (hasHOCTag(section)) {
+	if (await hasHOCTag(section)) {
 		return 'hoc';
 	}
 
 	// Check for @ui
-	if (hasComponentTag(section)) {
+	if (await hasComponentTag(section)) {
 		return 'component';
 	}